Changed sprintf to snprint, added ADC calibration

+++ Core/Inc/BMS_IC.h
@@ -11,13 +11,13 @@
 //Cell info buffers
 typedef struct batt_info{
 	float voltage_buffer[NUM_CELLS];
-	uint16_t temp_buffer[NUM_THEMISTOR];
-	uint16_t cell_volt_lowest;
-	uint16_t cell_volt_highest;
-	uint16_t cell_volt_avg;
-	uint16_t cell_volt_diff;
-	uint16_t cell_volt_sum;
-	uint16_t temp_avg;
+	float temp_buffer[NUM_THEMISTOR];
+	float cell_volt_lowest;
+	float cell_volt_highest;
+	float cell_volt_avg;
+	float cell_volt_diff;
+	float cell_volt_sum;
+	float temp_avg;
 }batt_info;
 
 void bms_ic_host_control_EN();
